♻️ refactor: 애매한 도메인 경로 이동 및 제거
