+++ src/renderer/src/features/champion/ChampionProfileSmall/ChampionProfileSmall.styled.ts
