+++ src/renderer/src/features/champion/ChampionProfileSmall/index.ts
