+++ src/renderer/src/features/champion/ChampionStats/ChampionStats.tsx
@@ -12,13 +12,13 @@ import { RuneIdsGroupByType } from '@main/modules/league/types/rune.types'
 import { CounterChampionsItem } from '@main/modules/league/types/stat.types'
 
 import LoadingIcon from '@renderer/components/LoadingIcon'
-import TierIcon, { HoneyIcon, OpIcon } from '@renderer/features/asset/TierIcon'
 import ItemBuildGroups from '@renderer/features/item/ItemBuildGroups'
 import LaneSelect from '@renderer/features/lane/LaneSelect'
 import RankRangeSelect from '@renderer/features/rank/RankRangeSelect'
 import { rankRangeIdAtom } from '@renderer/features/rank/RankRangeSelect/atoms/rankRangeId.atom'
 import RuneBuildButtonRadioList from '@renderer/features/rune/RuneBuildButtonRadioList'
 import RunePage from '@renderer/features/rune/RunePage'
+import TierIcon, { HoneyIcon, OpIcon } from '@renderer/features/tier/TierIcon'
 import useAPI from '@renderer/hooks/useAPI'
 import useCustomForm from '@renderer/hooks/useCustomForm'
