--- src/renderer/src/features/empty/LiveChampionSelectNotFound/LiveChampionSelectNotFound.styled.ts
@@ -1,3 +1,0 @@
-import styled from 'styled-components'
-
-export const Root = styled.div``
