--- src/renderer/src/features/empty/LiveChampionSelectNotFound/index.ts
@@ -1,3 +1,0 @@
-// === Automatically generated file. Don't edit it. ===
-export * from './LiveChampionSelectNotFound';
-export { default } from './LiveChampionSelectNotFound';
