--- src/renderer/src/features/inGame/InGameNotFound/InGameNotFound.styled.ts
@@ -1,3 +1,0 @@
-import styled from 'styled-components'
-
-export const Root = styled.div``
