--- src/renderer/src/features/inGame/InGameNotFound/index.ts
@@ -1,3 +1,0 @@
-// === Automatically generated file. Don't edit it. ===
-export * from './InGameNotFound';
-export { default } from './InGameNotFound';
