+++ src/renderer/src/features/rank/RankIcon/RankIcon.styled.ts
