+++ src/renderer/src/features/rank/RankIcon/RankIcon.tsx
