+++ src/renderer/src/features/rank/RankIcon/index.ts
