--- src/renderer/src/features/summoner/SummonerNotFound/SummonerNotFound.styled.ts
@@ -1,3 +1,0 @@
-import styled from 'styled-components'
-
-export const Root = styled.div``
