--- src/renderer/src/features/summoner/SummonerNotFound/index.ts
@@ -1,3 +1,0 @@
-// === Automatically generated file. Don't edit it. ===
-export * from './SummonerNotFound';
-export { default } from './SummonerNotFound';
