+++ src/renderer/src/features/tier/TierIcon/TierIcon.styled.ts
