+++ src/renderer/src/features/tier/TierIcon/TierIcon.tsx
