+++ src/renderer/src/features/tier/TierIcon/index.ts
