+++ src/renderer/src/pages/live/champion-select.tsx
@@ -1,8 +1,8 @@
+import { Result } from 'antd'
 import { useAtomValue } from 'jotai'
 
 import LayoutConfig from '@renderer/components/LayoutConfig'
 import ChampionStats from '@renderer/features/champion/ChampionStats'
-import LiveChampionSelectNotFound from '@renderer/features/empty/LiveChampionSelectNotFound'
 import { championSelectSessionAtom } from '@renderer/stores/atoms/championSelectSession.atom'
 import { configAtom } from '@renderer/stores/atoms/config.atom'
 import * as Styled from '@renderer/styles/pageStyled/LiveChampionSelectPage.styled'
@@ -17,7 +17,14 @@ const LiveChampionSelectPage = () => {
       <LayoutConfig breadcrumbs={['라이브 게임', '챔피언 선택']} />
 
       {hasChampionId && <LiveChampionDetail />}
-      {!hasChampionId && <LiveChampionSelectNotFound />}
+
+      {!hasChampionId && (
+        <Result
+          status="warning"
+          title="챔피언 선택 정보를 불러올 수 없습니다."
+          extra={<>현재 게임에 참여중이지 않거나, 챔피언 선택 정보를 불러올 수 없는 상태입니다.</>}
+        />
+      )}
     </Styled.Root>
   )
 }
